+++ src/index.js
@@ -31,3 +31,7 @@ export const map = _.curry((f, x) => (_.isArray(x) ? _.map : _.mapValues)(f, x))
 // defaults in multidimensional arrays and recursive plain objects
 export const deepMap = _.curry((fn, obj, _map = map, is = isTraversable) =>
   _map(e => is(e) ? deepMap(fn, fn(e), _map, is) : e, obj))
+
+
+// Lang
+export let throws = x => { throw x }