+++ src/index.js
@@ -4,6 +4,7 @@ export * from './conversion'
 export * from './function'
 export * from './array'
 export * from './object'
+export * from './aspect'
 export * from './regex'
 export * from './lens'
 
@@ -30,3 +31,6 @@ export const map = _.curry((f, x) => (_.isArray(x) ? _.map : _.mapValues)(f, x))
 // defaults in multidimensional arrays and recursive plain objects
 export const deepMap = _.curry((fn, obj, _map = map, is = isTraversable) =>
   _map(e => is(e) ? deepMap(fn, fn(e), _map, is) : e, obj))
+
+// Lang
+export let throws = x => { throw x }